backend/requirements.txt
fastapi>=0.111.0
uvicorn>=0.30.0
sqlmodel>=0.0.19
PyMySQL>=1.1.1
cryptography>=42.0.8
openai>=1.35.0
pydantic>=2.7.0
python-dotenv>=1.0.1
numpy>=1.26.4
